put srl gallery components up at index level
