+++ components/Gallery/Gallery.js
@@ -4,37 +4,33 @@ import ImageWithDescription from "./ImageDesc";
 import { VStack } from "@chakra-ui/react";
 
 export default function GalleryComponent(props) {
-  const options = {
-    settings: {
-      overlayColor: "rgba(30, 30, 30, 1)",
-    },
-    buttons: {
-      showDownloadButton: false,
-      showAutoplayButton: false,
-      showThumbnailsButton: false,
-    },
+  // const options = {
+  //   settings: {
+  //     overlayColor: "rgba(30, 30, 30, 1)",
+  //   },
+  //   buttons: {
+  //     showDownloadButton: false,
+  //     showAutoplayButton: false,
+  //     showThumbnailsButton: false,
+  //   },
 
-    thumbnails: {
-      showThumbnails: false,
-    },
-    caption: {
-      showCaption: false,
-    },
-  };
+  //   thumbnails: {
+  //     showThumbnails: false,
+  //   },
+  //   caption: {
+  //     showCaption: false,
+  //   },
+  // };
 
   const normLayout = props.layout !== "single";
   return (
-    <SimpleReactLightbox>
-      <SRLWrapper options={options}>
-        <VStack
-          spacing={normLayout && "10rem"}
-          align="stretch"
-          mt={{ base: "0rem", md: normLayout && "3rem" }}
-        >
-          <GallerySwitch {...props} />
-        </VStack>
-      </SRLWrapper>
-    </SimpleReactLightbox>
+    <VStack
+      spacing={normLayout && "10rem"}
+      align="stretch"
+      mt={{ base: "0rem", md: normLayout && "3rem" }}
+    >
+      <GallerySwitch {...props} />
+    </VStack>
   );
 }
